refactor: move models into src folder #591
