+++ server/src/api/controllers/cards.ts
@@ -2,10 +2,10 @@ import { Request, Response } from "express";
 import _ from "lodash";
 import { Types } from "mongoose";
 
-import { ICard } from "@models/User";
-import { Transaction } from "@models/Transaction";
+import { ICard } from "src/models/User";
+import { Transaction } from "src/models/Transaction";
 
-import { logger } from "src/loggers";
+import { logger } from "src/api/middleware/loggers";
 import { FIELD_REQUIRED, NOT_FOUND } from "../message";
 
 export const create = async (req: Request, res: Response) => {
