+++ server/src/api/controllers/categories.ts
@@ -1,12 +1,12 @@
 import { Request, Response } from "express";
 import _ from "lodash";
 import { Types } from "mongoose";
-import { Budget } from "@models/Budget";
-import { Transaction } from "@models/Transaction";
+import { Budget } from "src/models/Budget";
+import { Transaction } from "src/models/Transaction";
 
-import { ICategory } from "@models/User";
+import { ICategory } from "src/models/User";
 
-import { logger } from "src/loggers";
+import { logger } from "src/api/middleware/loggers";
 import {
   FIELD_INVALID,
   FIELD_REQUIRED,
