+++ server/src/api/controllers/test.ts
@@ -1,5 +1,5 @@
 import { Request, Response } from "express";
-import { logger } from "src/loggers";
+import { logger } from "src/api/middleware/loggers";
 
 /**
  * Hello
