+++ server/src/api/middleware/passport/kakaoStrategy.ts
@@ -1,8 +1,8 @@
 import passport from "passport";
 import { Strategy as KakaoStrategy, Profile } from "passport-kakao";
-import { User } from "@models/User";
+import { User } from "src/models/User";
 import { Request } from "express";
-import { CONNECTED_ALREADY, EMAIL_IN_USE, SNSID_IN_USE } from "../api/message";
+import { CONNECTED_ALREADY, EMAIL_IN_USE, SNSID_IN_USE } from "src/api/message";
 
 const getEmail = (profile: Profile): string | undefined => {
   return profile._json.kakao_account?.email;
