+++ server/src/models/Challenge.ts
