+++ server/src/models/_basicCategories.ts
