+++ server/src/models/_basicSettings.ts
