+++ server/tests/transactions.test.ts
@@ -5,7 +5,7 @@ import httpMocks from "node-mocks-http";
 import { Types } from "mongoose";
 
 jest.mock("@models/Transaction");
-import { Transaction } from "@models/Transaction";
+import { Transaction } from "src/models/Transaction";
 import * as transactions from "src/api/controllers/transactions";
 
 describe("GET /transactions?linkedPaymentMethodId=*** (결제 수단으로 카드 찾는 API)", () => {
